requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.2
Flask-Cors==3.0.10
Werkzeug==2.3.7
pytest==7.4.3
PyJWT==2.8.0
python-dotenv==1.0.0